# bff/api-docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Backend For Frontend API",
    "description": "API documentation for the BFF service.",
    "license": {
      "name": ""
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/users/{host_user_id}/watch-rooms": {
      "get": {
        "tags": [
          "WatchRooms"
        ],
        "operationId": "list_rooms_by_host",
        "parameters": [
          {
            "name": "host_user_id",
            "in": "path",
            "description": "UUID of the user hosting the rooms",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Max number of rooms to return",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 50
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Pagination offset",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            },
            "example": 0
          }
        ],
        "responses": {
          "200": {
            "description": "List rooms hosted by user",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/WatchRoom"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
    "/watch-rooms": {
      "post": {
        "tags": [
          "WatchRooms"
        ],
        "operationId": "create_watch_room",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NewWatchRoom"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Create a new watch room",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchRoom"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/watch-rooms/{room_id}": {
      "get": {
        "tags": [
          "WatchRooms"
        ],
        "operationId": "get_watch_room_by_id",
        "parameters": [
          {
            "name": "room_id",
            "in": "path",
            "description": "UUID of the watch room to retrieve",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Get watch room by ID",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchRoom"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "WatchRooms"
        ],
        "operationId": "update_watch_room",
        "parameters": [
          {
            "name": "room_id",
            "in": "path",
            "description": "UUID of the room to update",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWatchRoom"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Update an existing watch room",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WatchRoom"
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "delete": {
        "tags": [
          "WatchRooms"
        ],
        "operationId": "delete_watch_room",
        "parameters": [
          {
            "name": "room_id",
            "in": "path",
            "description": "UUID of the room to delete",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Delete a watch room",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "integer",
                  "minimum": 0
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/watch-rooms/{room_id}/connect": {
      "get": {
        "tags": [
          "WatchRooms"
        ],
        "operationId": "connect_room",
        "parameters": [
          {
            "name": "room_id",
            "in": "path",
            "description": "UUID of the room to connect to",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Connect to a watch room"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "NewWatchRoom": {
        "type": "object",
        "required": [
          "host_user_id",
          "content_id"
        ],
        "properties": {
          "content_id": {
            "type": "string",
            "format": "uuid"
          },
          "host_user_id": {
            "type": "string",
            "format": "uuid"
          },
          "invite_code": {
            "type": [
              "string",
              "null"
            ]
          },
          "is_private": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "room_name": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "UpdateWatchRoom": {
        "type": "object",
        "properties": {
          "current_time_seconds": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "is_live": {
            "type": [
              "boolean",
              "null"
            ]
          }
        }
      },
...
      "WatchRoom": {
        "type": "object",
        "required": [
          "id",
          "host_user_id",
          "content_id",
          "is_private",
          "created_at"
        ],
        "properties": {
          "content_id": {
            "type": "string",
            "format": "uuid"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "current_time_seconds": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32"
          },
          "host_user_id": {
            "type": "string",
            "format": "uuid"
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "invite_code": {
            "type": [
              "string",
              "null"
            ]
          },
          "is_live": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "is_private": {
            "type": "boolean"
          },
          "room_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "updated_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        }
      },
...
    }
  }
}